requirements.txt
psycopg2-binary==2.9.9
pydantic==2.9.1
python-dotenv==1.0.1
SQLAlchemy==1.4.54
apache-airflow[postgres]==2.10.5
